refactor: moved runner util to workflow
